+++ src/api/services/index.ts
@@ -3,3 +3,4 @@ export * from './userApi';
 export * from './invoiceApi';
 export * from './customerApi';
 export * from './serviceApi';
+export * from './reportsApi';