+++ src/constants/forms/formikInitialValues.ts
@@ -1,12 +1,11 @@
 import {
-  ArAgingInitialValueProps,
   ChangePasswordInitialValueProps,
   ForgetPwdProps,
   GstTypeProps,
   InvoiceInitialValueProps,
-  InvoicesInitialValueProps,
   LoginProps,
   PaymentTermsProps,
+  ReportsValueProps,
   RoleInitialValueProps,
   SendEmailInitialValueProps,
   SuperAdminUsersInitialValueProps,
@@ -132,14 +131,14 @@ export const invoiceCreateInitialValue: InvoiceInitialValueProps = {
   signatureFile: null,
 };
 
-export const invoicesInitialValue: InvoicesInitialValueProps = {
+export const invoicesInitialValue: ReportsValueProps = {
   // invoiceDate: "",
   startDate: '',
   endDate: '',
   filter: 'invoiceReport',
 };
 
-export const AragingInitialValue: ArAgingInitialValueProps = {
+export const AragingInitialValue: ReportsValueProps = {
   // invoiceDate: "",
   startDate: '',
   endDate: '',