+++ src/pages/company-users/UserForm.tsx
@@ -1,51 +1,42 @@
-import React, { useEffect, useMemo } from 'react';
+import React, { useMemo } from 'react';
 import { useSnackbarNotifications } from '../../hooks/useSnackbarNotification';
-import { clearData } from '../../redux-store/global/globalState';
-import { useDispatch } from 'react-redux';
-import { AppDispatch } from '../../app/store';
-import {
-  useCreateUserMutation,
-  useUpdateUserMutation,
-} from '../../redux-store/api/injectedApis';
 import { AdminCompanyUsersInitialValueProps } from '../../types/types';
-import { RoleValidationSchema } from '../../constants/forms/validations/validationSchema';
+import {
+  EditRoleValidationSchema,
+  RoleValidationSchema,
+} from '../../constants/forms/validations/validationSchema';
 import { DynamicFormCreate } from '../../components/Form-renderer/Dynamic-form';
 import {
   RolesEditFields,
   RolesFields,
 } from '../../constants/form-data/form-data-json';
 import { RoleInitialValue } from '../../constants/forms/formikInitialValues';
+import { useMutation } from '@tanstack/react-query';
+import { createUser, updateUser } from '../../api/services';
 
 interface UserValueProps {
-  userEditValue: any;
+  userEditValue?: any;
   mode: 'create' | 'edit';
   onClose: () => void;
-  refetchUserList: () => void;
 }
 
-const UserForm = ({
-  userEditValue,
-  mode,
-  onClose,
-  refetchUserList,
-}: UserValueProps) => {
-  const [
-    addUser,
-    {
-      isSuccess: userAddSuccess,
-      isError: userAddError,
-      error: userAddErrorObject,
+const UserForm = ({ userEditValue, mode, onClose }: UserValueProps) => {
+  const createUserMutation = useMutation({
+    mutationFn: createUser,
+    onSuccess: () => {
+      onClose();
     },
-  ] = useCreateUserMutation();
-  const [
-    updateUser,
-    {
-      isSuccess: userUpdateSuccess,
-      isError: userUpdateError,
-      error: userUpdateErrorObject,
+  });
+
+  const updateUserMutation = useMutation({
+    mutationFn: updateUser,
+    onSuccess: () => {
+      onClose();
     },
-  ] = useUpdateUserMutation();
-  const dispatch = useDispatch<AppDispatch>();
+  });
+
+  const isSuccess = createUserMutation.isSuccess;
+  const isError = createUserMutation.isError;
 
   // Setting initial values
   const initialValues =
@@ -53,39 +44,28 @@ const UserForm = ({
 
   // Snackbar Notifications
   useSnackbarNotifications({
-    error: userAddError,
-    errorObject: userAddErrorObject,
+    error: isError,
     errorMessage: 'Error creating user',
-    success: userAddSuccess,
-    successMessage: 'User created successfully',
+    success: isSuccess,
+    successMessage:
+      mode === 'edit'
+        ? 'User updated successfully'
+        : 'User created successfully',
   });
-
-  useSnackbarNotifications({
-    error: userUpdateError,
-    errorObject: userUpdateErrorObject,
-    errorMessage: 'Error updating user',
-    success: userUpdateSuccess,
-    successMessage: 'User updated successfully',
-  });
-
-  // Refetch data and close dialog on success
-  useEffect(() => {
-    if (userAddSuccess || userUpdateSuccess) {
-      refetchUserList();
-      onClose();
-    }
-  }, [userAddSuccess, userUpdateSuccess, refetchUserList, onClose]);
 
   // Form submission handler
   const onSubmit = useMemo(() => {
     return async (values: AdminCompanyUsersInitialValueProps, actions: any) => {
+      console.log('Form submitted with values:', values);
       try {
         if (mode === 'edit' && userEditValue) {
-          await updateUser({ id: values.id, data: { userDetails: values } });
+          console.log(values.id);
+          if (values.id) {
+            updateUserMutation.mutate({ id: values.id, data: values });
+          }
         } else {
-          await addUser({ userDetails: values });
+          createUserMutation.mutate({ userDetails: values });
         }
-        dispatch(clearData());
         actions.resetForm();
       } catch (error) {
         console.error('Error during form submission:', error);
@@ -93,7 +73,7 @@ const UserForm = ({
         actions.setSubmitting(false);
       }
     };
-  }, [updateUser, addUser, userEditValue, mode, dispatch]);
+  }, [userEditValue, mode]);
 
   return (
     <DynamicFormCreate
@@ -101,7 +81,9 @@ const UserForm = ({
       headerName={mode === 'create' ? 'User Create' : 'User Edit'}
       fields={mode === 'create' ? RolesFields : RolesEditFields}
       initialValues={initialValues}
-      validationSchema={RoleValidationSchema}
+      validationSchema={
+        mode === 'edit' ? EditRoleValidationSchema : RoleValidationSchema
+      }
       onSubmit={onSubmit}
     />
   );